requirements.txt
django
weasyprint
psycopg2
django-reversion
